# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "aptpay App API",
    "description": "Find out how your APIs work",
    "license": {
      "name": "tejas"
    }
  },
  "host": "uataptpay.moreyeahs.in",
  "basePath": "/api",
...
  "schemes": ["https", "http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "security": [
    {
      "APIKeyHeader": []
    }
  ],
  "paths": {
...
    "/merchant/getMerchantDetails": {
      "get": {
        "summary": "getMerchantDetails",
        "tags": ["merchant"],
        "description": "getMerchantDetails",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getMerchantDetailsResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getMerchantDetailsErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/getMerchantBalance": {
      "get": {
        "summary": "getMerchantBalance",
        "tags": ["merchant"],
        "description": "getMerchantBalance",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getMerchantBalanceResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getMerchantBalanceErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/getTransactionReport": {
      "get": {
        "summary": "getTransactionReport",
        "tags": ["merchant"],
        "description": "getTransactionReport",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getTransactionReportResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getTransactionReportErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/getFundingReport": {
      "get": {
        "summary": "getFundingReport",
        "tags": ["merchant"],
        "description": "getFundingReport",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getFundingReportResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getFundingReportErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/balanceToPayee": {
      "post": {
        "summary": "balanceToPayee",
        "tags": ["merchant"],
        "description": "balance To Payee",
        "parameters": [
          {
            "name": "balance To Payee",
            "in": "body",
            "description": "balanceToPayee",
            "schema": {
              "$ref": "#/definitions/balanceToPayee"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/balanceToPayeeResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/balanceToPayeeErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/getPayeeBalance": {
      "get": {
        "summary": "getPayeeBalance",
        "tags": ["merchant"],
        "description": "getPayeeBalance",
        "parameters": [
          {
            "name": "payeeId",
            "in": "query",
            "description": "getPayeeBalance",
            "schema": {
              "type": "number"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getPayeeBalanceResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getPayeeBalanceErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/getlimits": {
      "get": {
        "summary": "getlimitsDetails",
        "tags": ["merchant"],
        "description": "getlimits",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getlimitsResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getlimitsErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/getPrograms": {
      "get": {
        "summary": "getlimitsDetails",
        "tags": ["merchant"],
        "description": "getlimits",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getProgramsResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getProgramsErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/:merchantId/list": {
      "get": {
        "summary": "getMerchantList",
        "tags": ["merchant"],
        "description": "getMerchantList",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getMerchantListResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getMerchantListErrorResponse"
            }
          }
        }
      }
    },
    "/merchant/:program/age-restriction/:country": {
      "get": {
        "summary": "getProgramAgeRestriction",
        "tags": ["merchant"],
        "description": "getProgramAgeRestriction",
        "parameters": [
          {
            "name": "getProgramAgeRestriction",
            "in": "path",
            "schema": {
              "type": "object",
              "properties": {
                "program": {
                  "type": "string"
                },
                "country": {
                  "type": "string"
                }
              }
            },
            "required": true,
            "description": "Please enter _id"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/getProgramAgeRestrictionResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/getProgramAgeRestrictionErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "getMerchantDetailsResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getMerchantDetailsErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getMerchantBalanceResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getMerchantBalanceErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getTransactionReportResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getTransactionReportErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getFundingReportResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getFundingReportErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "balanceToPayee": {
      "properties": {
        "payeeId": {
          "type": "number"
        },
        "amount": {
          "type": "number"
        }
      }
    },
    "balanceToPayeeResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "properties": {
            "payeeId": {
              "type": "number"
            },
            "amount": {
              "type": "number"
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "balanceToPayeeErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getPayeeBalanceResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getPayeeBalanceErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getlimitsResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getlimitsErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getProgramsResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getProgramsErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getMerchantListResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getMerchantListErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  }
}